backend/requirements.txt
django
djangorestframework
mysqlclient
redis
